+++ clear_nas_access_u1.pas
@@ -13,13 +13,12 @@ type
   { TForm1 }
 
   TForm1 = class(TForm)
-    Button1: TButton;
-    Button2: TButton;
-    procedure Button1Click(Sender: TObject);
-    procedure Button2Click(Sender: TObject);
+    procedure FormClose(Sender: TObject; var CloseAction: TCloseAction);
+    procedure FormShow(Sender: TObject);
   private
+    plinkStarted : Boolean;
     SEInfo: TShellExecuteInfo;
-    function PlinkStarten : Integer;
+    function PlinkStarten : Boolean;
   public
 
   end;
@@ -33,54 +32,38 @@ implementation
 
 { TForm1 }
 
-procedure TForm1.Button1Click(Sender: TObject);
+procedure TForm1.FormClose(Sender: TObject; var CloseAction: TCloseAction);
 begin
-  PlinkStarten;
+  if plinkStarted then begin;
+    TerminateProcess(SEInfo.hProcess, 1);
+  end;
 end;
 
-procedure TForm1.Button2Click(Sender: TObject);
+procedure TForm1.FormShow(Sender: TObject);
 begin
-  TerminateProcess(SEInfo.hProcess, 1);
+  plinkStarted:= PlinkStarten;
 end;
 
-function TForm1.PlinkStarten: Integer;
+function TForm1.PlinkStarten: Boolean;
 var
   ExitCode: DWORD;
   ExecuteFile, ParamString, StartInString: string;
 begin
-  //ShellExecute(Handle, nil, 'putty.exe', nil, '.', SW_SHOWNORMAL);
-  //ExecuteFile:='putty.exe';
+  // https://www.thoughtco.com/execute-and-run-applications-1058462
   ExecuteFile:='plink.exe';
 
- FillChar(SEInfo, SizeOf(SEInfo), 0) ;
- SEInfo.cbSize := SizeOf(TShellExecuteInfo) ;
- with SEInfo do begin
- fMask := SEE_MASK_NOCLOSEPROCESS;
- Wnd := Application.Handle;
- lpFile := PChar(ExecuteFile) ;
-{
-ParamString can contain the
-application parameters.
-}
- lpParameters := PChar('-L 9999:192.168.42.2:80 -N pi@192.168.2.43 -pw kaas1234') ;
-{
-StartInString specifies the
-name of the working directory.
-If ommited, the current directory is used.
-}
- lpDirectory := PChar('.') ;
- nShow := SW_HIDE;
- end;
- ShellExecuteExA(@SEInfo)
- {if ShellExecuteExA(@SEInfo) then begin
- repeat
- Application.ProcessMessages;
- GetExitCodeProcess(SEInfo.hProcess, ExitCode) ;
- until (ExitCode <> STILL_ACTIVE) or
- Application.Terminated;
- ShowMessage('Calculator terminated') ;
- end
- else ShowMessage('Error starting Calc!') ;}
+  FillChar(SEInfo, SizeOf(SEInfo), 0) ;
+  SEInfo.cbSize := SizeOf(TShellExecuteInfo) ;
+  with SEInfo do begin
+    fMask := SEE_MASK_NOCLOSEPROCESS;
+    Wnd := Application.Handle;
+    lpFile := PChar(ExecuteFile) ;
+    lpParameters := PChar('-L 9999:192.168.42.2:80 -N pi@192.168.2.43 -pw kaas1234') ;
+
+    lpDirectory := PChar('.') ;
+    nShow := SW_HIDE;
+  end;
+  Result:= ShellExecuteExA(@SEInfo);
 end;
 
 end.
